refactor: modify design pattern to atomic
