--- src/ui/Pallete/index.ts
@@ -1,1 +1,0 @@
-export * from './Pallete';
