--- src/ui/ShareBox/index.ts
@@ -1,1 +1,0 @@
-export { default as ShareBox } from './ShareBox';
