+++ src/ui/atoms/BlinkIcon/index.ts
@@ -1,0 +1,1 @@
+export * from './BlinkIcon';
