+++ src/ui/atoms/index.ts
@@ -1,2 +1,4 @@
 export * from './Button';
 export * from './CardFrame';
+export * from './Input';
+export * from './BlinkIcon';
