+++ src/ui/molecules/Footer/index.ts
@@ -1,0 +1,1 @@
+export * from './Footer';
