+++ src/ui/molecules/Pallete/index.ts
