+++ src/ui/molecules/index.ts
@@ -1,0 +1,5 @@
+export * from './ResultCard';
+export * from './Pallete';
+export * from './ShareBox';
+export * from './Footer';
+export * from './Header';
