+++ src/ui/pages/index.ts
@@ -1,0 +1,6 @@
+export * from './FeedPage';
+export * from './IntroPage';
+export * from './Loading';
+export * from './Main';
+export * from './Result';
+export * from './WriteCard';
